fix all warnings under linux, enable all warnings (#28)
fixes all warnings on GCC 7.3.0. Not tested
using any other compile yet, so warnings should
still be expected using different compilers.

enables all warnings for all platforms, activates
'-pedantic' mode as well as updates the c++ standard
flag for non windows compilers.

note: This commit makes the compiler ignore
the '-Wunused-parameter' warning.

+++ interpreter.cpp
@@ -845,7 +845,7 @@ int32 interpreter::call_function(int32 f, uint32 param_count, int32 ret_count) {
 		return OPERATION_FAIL;
 	}
 	if (param_count != params.size()) {
-		sprintf(pduel->strbuffer, "\"CallFunction\": incorrect parameter count (%d expected, %ud pushed)", param_count, params.size());
+		sprintf(pduel->strbuffer, "\"CallFunction\": incorrect parameter count (%d expected, %ud pushed)", param_count, (uint32)params.size());
 		handle_message(pduel, 1);
 		params.clear();
 		return OPERATION_FAIL;
